+++ docker-compose.yml
@@ -27,7 +27,7 @@ services:
 
   transfair:
     container_name: transfair
-    image: samply/transfyr
+    image: samply/transfair
     environment:
       FHIR_INPUT_URL: "http://source_blaze:8080/fhir"
       FHIR_OUTPUT_URL: "http://target_blaze:8080/fhir"
