+++ pom.xml
@@ -11,8 +11,8 @@
 	<groupId>de.samply</groupId>
 	<artifactId>TransFAIR</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>transfyr</name>
-	<description>transfyr spring batch job</description>
+	<name>transFAIR</name>
+	<description>transFAIR spring batch job</description>
 	<properties>
 		<java.version>17</java.version>
 	</properties>
