+++ src/main/java/de/samply/transfair/TransFairApplication.java
@@ -1,0 +1,13 @@
+package de.samply.transfair;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TransFairApplication {
+
+	public static void main(String[] args) {
+		System.exit(SpringApplication.exit(SpringApplication.run(TransFairApplication.class, args)));
+	}
+
+}
