+++ src/main/java/de/samply/transfair/mapper/FhirResourceMapper.java
@@ -1,0 +1,10 @@
+package de.samply.transfair.mapper;
+
+import java.util.List;
+import org.hl7.fhir.r4.model.Resource;
+
+public interface FhirResourceMapper {
+  
+  public List<Resource> map(List<Resource> resources);
+
+}
