--- src/main/java/de/samply/transfyr/TransfyrApplication.java
@@ -1,13 +1,0 @@
-package de.samply.transfyr;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TransfyrApplication {
-
-	public static void main(String[] args) {
-		System.exit(SpringApplication.exit(SpringApplication.run(TransfyrApplication.class, args)));
-	}
-
-}
