+++ src/test/java/de/samply/transfair/TransFairApplicationTests.java
@@ -1,0 +1,13 @@
+package de.samply.transfair;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TransFairApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
