--- src/test/java/de/samply/transfyr/TransfyrApplicationTests.java
@@ -1,13 +1,0 @@
-package de.samply.transfyr;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TransfyrApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
